move kafka serialization wrappers to kafka.internals package
